+++ store/storage/sqlite/db.go
@@ -128,7 +128,7 @@ func (db *Database) SetLatestVersion(version uint64) error {
 	return nil
 }
 
-func (db *Database) Has(storeKey string, version uint64, key []byte) (bool, error) {
+func (db *Database) Has(storeKey []byte, version uint64, key []byte) (bool, error) {
 	val, err := db.Get(storeKey, version, key)
 	if err != nil {
 		return false, err
@@ -137,7 +137,7 @@ func (db *Database) Has(storeKey string, version uint64, key []byte) (bool, erro
 	return val != nil, nil
 }
 
-func (db *Database) Get(storeKey string, targetVersion uint64, key []byte) ([]byte, error) {
+func (db *Database) Get(storeKey []byte, targetVersion uint64, key []byte) ([]byte, error) {
 	if targetVersion < db.earliestVersion {
 		return nil, storeerrors.ErrVersionPruned{EarliestVersion: db.earliestVersion}
 	}
@@ -216,7 +216,7 @@ func (db *Database) Prune(version uint64) error {
 	return nil
 }
 
-func (db *Database) Iterator(storeKey string, version uint64, start, end []byte) (corestore.Iterator, error) {
+func (db *Database) Iterator(storeKey []byte, version uint64, start, end []byte) (corestore.Iterator, error) {
 	if (start != nil && len(start) == 0) || (end != nil && len(end) == 0) {
 		return nil, storeerrors.ErrKeyEmpty
 	}
@@ -228,7 +228,7 @@ func (db *Database) Iterator(storeKey string, version uint64, start, end []byte)
 	return newIterator(db, storeKey, version, start, end, false)
 }
 
-func (db *Database) ReverseIterator(storeKey string, version uint64, start, end []byte) (corestore.Iterator, error) {
+func (db *Database) ReverseIterator(storeKey []byte, version uint64, start, end []byte) (corestore.Iterator, error) {
 	if (start != nil && len(start) == 0) || (end != nil && len(end) == 0) {
 		return nil, storeerrors.ErrKeyEmpty
 	}
@@ -256,7 +256,7 @@ func (db *Database) PrintRowsDebug() {
 	var sb strings.Builder
 	for rows.Next() {
 		var (
-			storeKey string
+			storeKey []byte
 			key      []byte
 			value    []byte
 			version  uint64