+++ tests/integration/staking/keeper/genesis_test.go
@@ -15,6 +15,7 @@ import (
 
 	codectypes "github.com/cosmos/cosmos-sdk/codec/types"
 	sdk "github.com/cosmos/cosmos-sdk/types"
+	"github.com/cosmos/cosmos-sdk/types/module"
 )
 
 func bootstrapGenesisTest(t *testing.T, numAddrs int) (*fixture, []sdk.AccAddress) {
@@ -98,9 +99,11 @@ func TestInitGenesis(t *testing.T) {
 	delegations = append(delegations, genesisDelegations...)
 
 	genesisState := types.NewGenesisState(params, validators, delegations)
-	vals := (f.stakingKeeper.InitGenesis(f.sdkCtx, genesisState))
+	vals, err := (f.stakingKeeper.InitGenesis(f.sdkCtx, genesisState))
+	assert.NilError(t, err)
 
-	actualGenesis := (f.stakingKeeper.ExportGenesis(f.sdkCtx))
+	actualGenesis, err := (f.stakingKeeper.ExportGenesis(f.sdkCtx))
+	assert.NilError(t, err)
 	assert.DeepEqual(t, genesisState.Params, actualGenesis.Params)
 	assert.DeepEqual(t, genesisState.Delegations, actualGenesis.Delegations)
 
@@ -126,12 +129,11 @@ func TestInitGenesis(t *testing.T) {
 	assert.Equal(t, types.Bonded, resVal.Status)
 
 	abcivals := make([]abci.ValidatorUpdate, len(vals))
-
+	validatorUpdates := make([]module.ValidatorUpdate, len(abcivals))
 	for i, val := range validators {
-		abcivals[i] = val.ABCIValidatorUpdate((f.stakingKeeper.PowerReduction(f.sdkCtx)))
+		validatorUpdates[i] = val.ModuleValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx))
 	}
-
-	assert.DeepEqual(t, abcivals, vals)
+	assert.DeepEqual(t, validatorUpdates, vals)
 }
 
 func TestInitGenesis_PoolsBalanceMismatch(t *testing.T) {
@@ -157,27 +159,23 @@ func TestInitGenesis_PoolsBalanceMismatch(t *testing.T) {
 		BondDenom:     "stake",
 	}
 
-	require.Panics(t, func() {
-		// setting validator status to bonded so the balance counts towards bonded pool
-		validator.Status = types.Bonded
-		f.stakingKeeper.InitGenesis(f.sdkCtx, &types.GenesisState{
-			Params:     params,
-			Validators: []types.Validator{validator},
-		})
-	},
-	// "should panic because bonded pool balance is different from bonded pool coins",
-	)
+	// setting validator status to bonded so the balance counts towards bonded pool
+	validator.Status = types.Bonded
+	_, err = f.stakingKeeper.InitGenesis(f.sdkCtx, &types.GenesisState{
+		Params:     params,
+		Validators: []types.Validator{validator},
+	})
+	// "should error because bonded pool balance is different from bonded pool coins",
+	require.NotNil(t, err)
 
-	require.Panics(t, func() {
-		// setting validator status to unbonded so the balance counts towards not bonded pool
-		validator.Status = types.Unbonded
-		f.stakingKeeper.InitGenesis(f.sdkCtx, &types.GenesisState{
-			Params:     params,
-			Validators: []types.Validator{validator},
-		})
-	},
+	// setting validator status to unbonded so the balance counts towards not bonded pool
+	validator.Status = types.Unbonded
+	_, err = f.stakingKeeper.InitGenesis(f.sdkCtx, &types.GenesisState{
+		Params:     params,
+		Validators: []types.Validator{validator},
+	})
 	// "should panic because not bonded pool balance is different from not bonded pool coins",
-	)
+	require.NotNil(t, err)
 }
 
 func TestInitGenesisLargeValidatorSet(t *testing.T) {
@@ -228,14 +226,14 @@ func TestInitGenesisLargeValidatorSet(t *testing.T) {
 		),
 	)
 
-	vals := f.stakingKeeper.InitGenesis(f.sdkCtx, genesisState)
+	vals, err := f.stakingKeeper.InitGenesis(f.sdkCtx, genesisState)
+	assert.NilError(t, err)
 
-	abcivals := make([]abci.ValidatorUpdate, 100)
+	validatorUpdates := make([]module.ValidatorUpdate, 100)
 	for i, val := range validators[:100] {
-		abcivals[i] = val.ABCIValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx))
+		validatorUpdates[i] = val.ModuleValidatorUpdate(f.stakingKeeper.PowerReduction(f.sdkCtx))
 	}
-
 	// remove genesis validator
 	vals = vals[:100]
-	assert.DeepEqual(t, abcivals, vals)
+	assert.DeepEqual(t, validatorUpdates, vals)
 }