+++ x/bank/keeper/genesis.go
@@ -57,10 +57,10 @@ func (k BaseKeeper) InitGenesis(ctx context.Context, genState *types.GenesisStat
 }
 
 // ExportGenesis returns the bank module's genesis state.
-func (k BaseKeeper) ExportGenesis(ctx context.Context) *types.GenesisState {
+func (k BaseKeeper) ExportGenesis(ctx context.Context) (*types.GenesisState, error) {
 	totalSupply, _, err := k.GetPaginatedTotalSupply(ctx, &query.PageRequest{Limit: query.PaginationMaxLimit})
 	if err != nil {
-		panic(fmt.Errorf("unable to fetch total supply %v", err))
+		return nil, fmt.Errorf("unable to fetch total supply %v", err)
 	}
 
 	rv := types.NewGenesisState(
@@ -70,5 +70,5 @@ func (k BaseKeeper) ExportGenesis(ctx context.Context) *types.GenesisState {
 		k.GetAllDenomMetaData(ctx),
 		k.GetAllSendEnabledEntries(ctx),
 	)
-	return rv
+	return rv, nil
 }