+++ x/genutil/module.go
@@ -5,8 +5,6 @@ import (
 	"encoding/json"
 	"fmt"
 
-	abci "github.com/cometbft/cometbft/abci/types"
-
 	"cosmossdk.io/core/appmodule"
 	"cosmossdk.io/core/genesis"
 	"cosmossdk.io/core/registry"
@@ -77,19 +75,15 @@ func (am AppModule) ValidateGenesis(bz json.RawMessage) error {
 }
 
 // InitGenesis performs genesis initialization for the genutil module.
-func (am AppModule) InitGenesis(ctx context.Context, data json.RawMessage) []abci.ValidatorUpdate {
+func (am AppModule) InitGenesis(ctx context.Context, data json.RawMessage) ([]module.ValidatorUpdate, error) {
 	var genesisState types.GenesisState
 	am.cdc.MustUnmarshalJSON(data, &genesisState)
-	validators, err := InitGenesis(ctx, am.stakingKeeper, am.deliverTx, genesisState, am.txEncodingConfig)
-	if err != nil {
-		panic(err)
-	}
-	return validators
+	return InitGenesis(ctx, am.stakingKeeper, am.deliverTx, genesisState, am.txEncodingConfig)
 }
 
 // ExportGenesis returns the exported genesis state as raw bytes for the genutil module.
-func (am AppModule) ExportGenesis(_ context.Context) json.RawMessage {
-	return am.DefaultGenesis()
+func (am AppModule) ExportGenesis(_ context.Context) (json.RawMessage, error) {
+	return am.DefaultGenesis(), nil
 }
 
 // GenTxValidator returns the genutil module's genesis transaction validator.
@@ -101,4 +95,4 @@ func (am AppModule) GenTxValidator() types.MessageValidator {
 func (AppModule) ConsensusVersion() uint64 { return 1 }
 
 // RegisterInterfaces implements module.AppModule.
-func (AppModule) RegisterInterfaces(registry.LegacyRegistry) {}
+func (AppModule) RegisterInterfaces(registry.InterfaceRegistrar) {}