+++ x/gov/keeper/proposal.go
@@ -146,7 +146,9 @@ func (k Keeper) SubmitProposal(ctx context.Context, messages []sdk.Msg, metadata
 
 	if err := k.environment.EventService.EventManager(ctx).EmitKV(
 		types.EventTypeSubmitProposal,
+		event.NewAttribute(types.AttributeKeyProposalType, proposalType.String()),
 		event.NewAttribute(types.AttributeKeyProposalID, fmt.Sprintf("%d", proposalID)),
+		event.NewAttribute(types.AttributeKeyProposalProposer, proposer.String()),
 		event.NewAttribute(types.AttributeKeyProposalMessages, strings.Join(msgs, ",")),
 	); err != nil {
 		return v1.Proposal{}, fmt.Errorf("failed to emit event: %w", err)